# packages/api-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ReCredita API",
    "description": "Recuperação (Plano A) e aquisição (Plano B) de veículos.",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {
...
      "Erro": {
        "type": "object",
        "properties": {
          "erro": {
            "type": "string"
          }
        },
        "required": [
          "erro"
        ],
        "additionalProperties": false
      },
...
      "ResumoImportacao": {
        "type": "object",
        "properties": {
          "relatorioId": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "veiculo": {
            "nullable": true,
            "type": "object",
            "properties": {
              "campos": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "restricoes": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "campos",
              "restricoes"
            ],
            "additionalProperties": false
          },
          "pessoas": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "nome": {
                  "nullable": true,
                  "type": "string"
                },
                "papeis": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "enum": [
                      "devedor",
                      "proprietario",
                      "terceiro_possuidor",
                      "parente",
                      "avalista",
                      "outro"
                    ]
                  }
                }
              },
              "required": [
                "id",
                "nome",
                "papeis"
              ],
              "additionalProperties": false
            }
          },
          "contatos": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "enderecos": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "parentes": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "radares": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "extras": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "novos": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "secao": {
                  "type": "string"
                },
                "rotulo": {
                  "type": "string"
                }
              },
              "required": [
                "secao",
                "rotulo"
              ],
              "additionalProperties": false
            }
          },
          "avisos": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "relatorioId",
          "veiculo",
          "pessoas",
          "contatos",
          "enderecos",
          "parentes",
          "radares",
          "extras",
          "novos",
          "avisos"
        ],
        "additionalProperties": false
      },
      "RelatorioColado": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "via": {
            "type": "string",
            "enum": [
              "colado",
              "integracao"
            ]
          },
          "secoes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "resumo": {
            "type": "object",
            "additionalProperties": {}
          },
          "fornecedor": {
            "nullable": true,
            "type": "string"
          },
          "usuarioNome": {
            "nullable": true,
            "type": "string"
          },
          "coladoEm": {
            "type": "string",
            "format": "date-time"
          },
          "tamanho": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "id",
          "via",
          "secoes",
          "resumo",
          "fornecedor",
          "usuarioNome",
          "coladoEm",
          "tamanho"
        ],
        "additionalProperties": false
      },
      "DadoExtra": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "entidade": {
            "type": "string",
            "enum": [
              "veiculo",
              "pessoa",
              "caso"
            ]
          },
          "pessoaId": {
            "nullable": true,
            "type": "string"
          },
          "secao": {
            "type": "string"
          },
          "chave": {
            "type": "string"
          },
          "rotulo": {
            "type": "string"
          },
          "valor": {
            "type": "string"
          },
          "sensivel": {
            "type": "boolean"
          },
          "novo": {
            "type": "boolean"
          },
          "promovidoEm": {
            "nullable": true,
            "type": "string",
            "format": "date-time"
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "entidade",
          "pessoaId",
          "secao",
          "chave",
          "rotulo",
          "valor",
          "sensivel",
          "novo",
          "promovidoEm",
          "criadoEm"
        ],
        "additionalProperties": false
      },
      "PessoasDoCaso": {
        "type": "object",
        "properties": {
          "alertas": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pessoas": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "papel": {
                  "type": "string",
                  "enum": [
                    "devedor",
                    "proprietario",
                    "terceiro_possuidor",
                    "parente",
                    "avalista",
                    "outro"
                  ]
                },
                "vinculo": {
                  "nullable": true,
                  "type": "string"
                },
                "iniciais": {
                  "nullable": true,
                  "type": "string"
                },
                "documentoMascarado": {
                  "nullable": true,
                  "type": "string"
                },
                "tipoPessoa": {
                  "nullable": true,
                  "type": "string",
                  "enum": [
                    "PF",
                    "PJ"
                  ]
                },
                "obito": {
                  "nullable": true,
                  "type": "boolean"
                },
                "contatos": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "enderecos": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "id",
                "papel",
                "vinculo",
                "iniciais",
                "documentoMascarado",
                "tipoPessoa",
                "obito",
                "contatos",
                "enderecos"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "alertas",
          "pessoas"
        ],
        "additionalProperties": false
      },
...
      "PessoasReveladas": {
        "type": "object",
        "properties": {
          "alertas": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sensivelLiberado": {
            "type": "boolean"
          },
          "pessoas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PessoaRevelada"
            }
          }
        },
        "required": [
          "alertas",
          "sensivelLiberado",
          "pessoas"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/api/casos/{id}/relatorios": {
      "post": {
        "tags": [
          "relatorios"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "texto": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 500000
                  },
                  "papelPessoa": {
                    "type": "string",
                    "enum": [
                      "devedor",
                      "proprietario",
                      "terceiro_possuidor",
                      "avalista",
                      "outro"
                    ]
                  },
                  "bureauId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "baseLegal": {
                    "type": "string",
                    "enum": [
                      "execucao_contrato",
                      "legitimo_interesse",
                      "obrigacao_legal",
                      "consentimento"
                    ]
                  },
                  "justificativa": {
                    "type": "string",
                    "minLength": 10
                  }
                },
                "required": [
                  "texto"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumoImportacao"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "relatorios"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RelatorioColado"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/casos/{id}/conferir-devedor": {
      "post": {
        "tags": [
          "relatorios"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "documentos": {
                    "maxItems": 20,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[0-9A-Z]{11,14}$"
                    }
                  }
                },
                "required": [
                  "documentos"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "iguais": {
                      "type": "array",
                      "items": {
                        "type": "boolean"
                      }
                    }
                  },
                  "required": [
                    "iguais"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/casos/{id}/relatorios/{relatorioId}/texto": {
      "post": {
        "tags": [
          "relatorios"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "finalidade": {
                    "type": "string",
                    "minLength": 10
                  }
                },
                "required": [
                  "finalidade"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "relatorioId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "texto": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "texto"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/casos/{id}/dados-extras": {
      "get": {
        "tags": [
          "relatorios"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DadoExtra"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/casos/{id}/pessoas": {
      "get": {
        "tags": [
          "relatorios"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PessoasDoCaso"
                }
              }
            }
          }
        }
      }
    },
    "/api/casos/{id}/pessoas/revelar": {
      "post": {
        "tags": [
          "relatorios"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "finalidade": {
                    "type": "string",
                    "minLength": 10
                  }
                },
                "required": [
                  "finalidade"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PessoasReveladas"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
...
  }
}